algo-trader-pro/src/execution/live_executor.py
"""
src/execution/live_executor.py

Live Trading Executor - Safety Stub.

This module intentionally raises NotImplementedError on instantiation to
prevent accidental live trading.  Real money execution requires explicit
implementation of authenticated Bybit API calls, order management, and
robust error handling (rate limits, partial fills, connectivity loss).

To enable live trading:
1. Implement all methods below using the pybit SDK.
2. Add API key / secret management via environment variables.
3. Thoroughly test on Bybit Testnet before switching to mainnet.
4. Set app.environment = 'live' in settings.yaml AND explicitly override
   the safety guard in this class after careful review.
"""

from __future__ import annotations

import logging


class LiveExecutor:
    """
    LIVE TRADING EXECUTOR - NOT YET ENABLED

    This is a safety stub.  Live trading requires explicit activation,
    API key configuration, testnet validation, and a conscious decision
    to risk real capital.

    Attempting to instantiate this class will raise NotImplementedError.

    How to implement (future work)
    ------------------------------
    1. Install pybit:  pip install pybit
    2. Authenticate:
           from pybit.unified_trading import HTTP
           self._client = HTTP(
               testnet=use_testnet,
               api_key=os.environ['BYBIT_API_KEY'],
               api_secret=os.environ['BYBIT_API_SECRET'],
           )
    3. Implement execute_buy using client.place_order(...)
    4. Implement execute_sell using client.place_order(side='Sell', ...)
    5. Implement check_stops by polling client.get_positions(...)
    6. Add WebSocket feed for real-time stop management.
    7. Handle PartialFill, Cancelled, and Rejected order states.
    8. Add circuit-breaker logic: stop trading on consecutive API errors.
    """

    def __init__(self, *args, **kwargs) -> None:
        _logger = logging.getLogger(__name__)
        _logger.critical(
            "\n"
            "=" * 70 + "\n"
            "  LIVE EXECUTOR INITIALIZED - THIS WILL USE REAL MONEY\n"
            "  Live trading is not yet enabled.\n"
            "  Use PaperExecutor for paper trading and backtesting.\n"
            "  To enable live trading, implement this class with Bybit API calls.\n"
            "=" * 70
        )
        raise NotImplementedError(
            "Live trading not enabled.\n"
            "Use PaperExecutor for paper trading.\n"
            "To enable live trading, implement this class with Bybit API calls\n"
            "and remove this guard after thorough testnet validation."
        )

    def execute_buy(self, *args, **kwargs):
        """Place a live market buy / entry order."""
        raise NotImplementedError("Live trading not enabled.")

    def execute_sell(self, *args, **kwargs):
        """Place a live market sell / exit order."""
        raise NotImplementedError("Live trading not enabled.")

    def check_stops(self, *args, **kwargs):
        """Check open positions against current market prices."""
        raise NotImplementedError("Live trading not enabled.")

    def update_trailing_stops(self, *args, **kwargs):
        """Update trailing stop levels on the exchange."""
        raise NotImplementedError("Live trading not enabled.")

    def get_portfolio_value(self, *args, **kwargs):
        """Query live account equity from the exchange."""
        raise NotImplementedError("Live trading not enabled.")

    def get_equity_snapshot(self, *args, **kwargs):
        """Return a snapshot of live account state."""
        raise NotImplementedError("Live trading not enabled.")

    def get_trade_stats(self, *args, **kwargs):
        """Return trade statistics from live account history."""
        raise NotImplementedError("Live trading not enabled.")
