algo-trader-pro/requirements.txt
# =============================================================================
# AlgoTrader Pro - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# Python >= 3.11 recommended

# -----------------------------------------------------------------------------
# Web Framework & API
# -----------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
websockets>=12.0

# -----------------------------------------------------------------------------
# Configuration & Environment
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.7.0
pydantic-settings>=2.0.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=16.0.0

# -----------------------------------------------------------------------------
# Technical Analysis (pandas-ta-classic: fork compatibile, wheel su PyPI)
# -----------------------------------------------------------------------------
pandas-ta-classic>=0.3.14b2

# -----------------------------------------------------------------------------
# HTTP Clients
# -----------------------------------------------------------------------------
requests>=2.31.0
aiohttp>=3.9.0

# -----------------------------------------------------------------------------
# Machine Learning
# -----------------------------------------------------------------------------
xgboost>=2.0.0
scikit-learn>=1.4.0
torch>=2.2.0
optuna>=3.6.0

# -----------------------------------------------------------------------------
# Sentiment Analysis
# -----------------------------------------------------------------------------
vaderSentiment>=3.3.2
pytrends>=4.9.2

# -----------------------------------------------------------------------------
# Task Scheduling
# -----------------------------------------------------------------------------
apscheduler>=3.10.0

# -----------------------------------------------------------------------------
# Development & Testing (optional but recommended)
# -----------------------------------------------------------------------------
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=5.0.0
# black>=24.0.0
# isort>=5.13.0
# ruff>=0.4.0
# mypy>=1.9.0
